+++ lib/default.nix
@@ -6,7 +6,7 @@ lib:
 with lib; # Android stuff that might come in handy across multiple modules, profiles, etc
 rec {
   # Guess resource type. Should normally work fine, but can't detect color/dimension types
-  resourceType = r:
+  resourceTypeName = r:
     if isBool r then "bool"
     else if isInt r then "integer"
     else if isString r then "string"
@@ -27,13 +27,21 @@ rec {
     string-array = concatMapStringsSep "" (i: "<item>${i}</item>") value;
     # Ignoring other typed arrays for now
   }.${type};
-  resourceXML = name: value: type: ''<${type} name="${name}">${resourceValueXML value type}</${type}>'';
+
+  resourceXML = name: value: let
+    resourceXMLEntity = name: value: type: ''<${type} name="${name}">${resourceValueXML value type}</${type}>'';
+  in if isAttrs value then
+      # Submodule with manually specified resource type
+      resourceXMLEntity name value.value value.type
+    else
+      # Bare value, so use Autodetected resource type
+      resourceXMLEntity name value (resourceTypeName value);
+
   configXML = resources: ''
     <?xml version="1.0" encoding="utf-8"?>
     <resources>
-      ${concatStringsSep "\n" (mapAttrsToList
-        (name: value: resourceXML name value (resourceType value))
-       resources)}
+      ${concatStringsSep "\n" (mapAttrsToList resourceXML resources)}
     </resources>
   '';
+
 }