+++ modules/resources.nix
@@ -7,28 +7,33 @@
 # https://developer.android.com/guide/topics/resources/more-resources.html
 with lib;
 let
-  # TODO: Unify with robotnixlib.configXML (just need something nice for resourceTypeOverrides)
-  configXML = relativePath: packageResources: ''
-    <?xml version="1.0" encoding="utf-8"?>
-    <resources>
-      ${concatStringsSep "\n" (mapAttrsToList
-        (name: value: robotnixlib.resourceXML name value (config.resourceTypeOverrides.${relativePath}.${name} or (robotnixlib.resourceType value)))
-       packageResources)}
-    </resources>
-  '';
+  # TODO:
+  # "oneOf [ (listOf int) (listOf str) ]" doesn't work. Fails with str looking for int
+  # Using "listOf (either str int)" instead
+  resourceTypeGeneric = with types; oneOf [ bool int str (listOf (either str int)) ];
+  resourceTypeModule = types.submodule ({ name, config, ... }: {
+    options = {
+      type = mkOption {
+        type = types.strMatching "(bool|integer|dimen|color|string|integer-array|string-array)";
+        description = "Set to override auto-detected resource type";
+      };
+
+      value = mkOption {
+        type = resourceTypeGeneric;
+      };
+    };
+
+    config = {
+      type = mkOptionDefault (resourceTypeName config.value);
+    };
+  });
 in
 {
   options = {
     resources = mkOption {
       default = {};
-      type = types.attrsOf (types.attrsOf types.unspecified);
-      description = "package resources. The first key refers to the relative path for the package, and the second key refers to the resource name";
-    };
-
-    resourceTypeOverrides = mkOption {
-      default = {};
-      type = types.attrsOf (types.attrsOf (types.strMatching "(bool|integer|dimen|color|string|integer-array|string-array)"));
-      description = "for overriding auto-detected resource type";
+      type = with types; attrsOf (attrsOf (either resourceTypeGeneric resourceTypeModule));
+      description = "Package resources. The first key refers to the relative path for the package, and the second key refers to the resource name";
     };
   };
 
@@ -40,7 +45,7 @@ in
       name = "robotnix-overlay";
       paths = mapAttrsToList (relativePath: packageResources: (pkgs.writeTextFile {
         name = "${replaceStrings ["/"] ["="] relativePath}-resources";
-        text = configXML relativePath packageResources;
+        text = robotnixlib.configXML packageResources;
         destination = "/${relativePath}/res/values/default.xml"; # I think it's ok that the name doesn't match the original--since they all get merged anyway
       })) config.resources;
     });