+++ NEWS.md
@@ -8,6 +8,11 @@ changes occur which require user intervention / configuration changes.  These
 are highlights since the last update, and are not meant to be an exhaustive
 listing of changes. See the git commit log for additional details.
 
+# 2021-03-XX
+
+## Backward incompatible changes
+- The `resourceTypeOverrides` option was replaced with `resources.<package>.<name>.type`.
+
 # 2021-02-02
 ## Highlights:
 - Updated vanilla flavor to RQ1A.210205.004